add imported query
