+++ wp-templates/page.js
@@ -17,6 +17,7 @@ import { useRouter } from "next/router";
 import  Link  from  'next/link';
 import { WordPressBlocksViewer } from '@faustwp/blocks';
 import { flatListToHierarchical } from '@faustwp/core';
+import GET_PAGE from '../queries/test';
  
 export default function Component(props) {
   const { locale: activeLocale, locales } = useRouter();
@@ -79,45 +80,47 @@ Component.variables = ({databaseId, language}, ctx) => {
   };
 };
 
-Component.query = gql`
-  ${BlogInfoFragment}
-  ${NavigationMenu.fragments.entry}
-  ${FeaturedImage.fragments.entry}
-  query GetPageData(
-    $databaseId: ID!
-    $headerLocation: MenuLocationEnum
-    $footerLocation: MenuLocationEnum
-    $asPreview: Boolean = false
-  ) {
-    page(id: $databaseId, idType: DATABASE_ID, asPreview: $asPreview) {
-      title
-      content
-      editorBlocks {
-        name
-        renderedHtml
-        clientId
-        parentClientId
-      }
-      translations {
-        uri
-        language {
-          code
-        }
-      }
-      ...FeaturedImageFragment
-    }
-    generalSettings {
-      ...BlogInfoFragment
-    }
-    footerMenuItems: menuItems(where: { location: $footerLocation }) {
-      nodes {
-        ...NavigationMenuItemFragment
-      }
-    }
-    headerMenuItems: menuItems(where: { location: $headerLocation }) {
-      nodes {
-        ...NavigationMenuItemFragment
-      }
-    }
-  }
-`;
+Component.query = GET_PAGE;
+
+// Component.query = gql`
+//   ${BlogInfoFragment}
+//   ${NavigationMenu.fragments.entry}
+//   ${FeaturedImage.fragments.entry}
+//   query GetPageData(
+//     $databaseId: ID!
+//     $headerLocation: MenuLocationEnum
+//     $footerLocation: MenuLocationEnum
+//     $asPreview: Boolean = false
+//   ) {
+//     page(id: $databaseId, idType: DATABASE_ID, asPreview: $asPreview) {
+//       title
+//       content
+//       editorBlocks {
+//         name
+//         renderedHtml
+//         clientId
+//         parentClientId
+//       }
+//       translations {
+//         uri
+//         language {
+//           code
+//         }
+//       }
+//       ...FeaturedImageFragment
+//     }
+//     generalSettings {
+//       ...BlogInfoFragment
+//     }
+//     footerMenuItems: menuItems(where: { location: $footerLocation }) {
+//       nodes {
+//         ...NavigationMenuItemFragment
+//       }
+//     }
+//     headerMenuItems: menuItems(where: { location: $headerLocation }) {
+//       nodes {
+//         ...NavigationMenuItemFragment
+//       }
+//     }
+//   }
+// `;
